Match perf_fast_untilize with functional test counterpart.
